+++ nestjs/src/app.module.ts
@@ -20,7 +20,7 @@ import { CertificatesModule } from './certificates/certificates.module';
 import { DiscordServersModule } from './discord-servers/discord-servers.module';
 import { CrossCheckModule } from './cross-check/cross-check.module';
 import { OpportunitiesModule } from './opportunities/opportunities.module';
-import { UserGroupsModule } from './userGroups/user-groups.module';
+import { UserGroupsModule } from './user-groups/user-groups.module';
 import { ScheduleModule } from './schedule/schedule.module';
 import { ActivityModule } from './activity/activity.module';
 import { UsersNotificationsModule } from './users-notifications';
